run-prettier

+++ packages/playwright-core/src/client/channelOwner.ts
@@ -25,7 +25,9 @@ import { ClientInstrumentation } from './clientInstrumentation';
 import type { Connection } from './connection';
 import type { Logger } from './types';
 
-export abstract class ChannelOwner<T extends channels.Channel = channels.Channel> extends EventEmitter {
+export abstract class ChannelOwner<
+  T extends channels.Channel = channels.Channel,
+> extends EventEmitter {
   readonly _connection: Connection;
   private _parent: ChannelOwner | undefined;
   private _objects = new Map<string, ChannelOwner>();
@@ -37,7 +39,13 @@ export abstract class ChannelOwner<T extends channels.Channel = channels.Channel
   _logger: Logger | undefined;
   _instrumentation: ClientInstrumentation | undefined;
 
-  constructor(parent: ChannelOwner | Connection, type: string, guid: string, initializer: channels.InitializerTraits<T>, instrumentation?: ClientInstrumentation) {
+  constructor(
+    parent: ChannelOwner | Connection,
+    type: string,
+    guid: string,
+    initializer: channels.InitializerTraits<T>,
+    instrumentation?: ClientInstrumentation,
+  ) {
     super();
     this.setMaxListeners(0);
     this._connection = parent instanceof ChannelOwner ? parent._connection : parent;
@@ -58,20 +66,18 @@ export abstract class ChannelOwner<T extends channels.Channel = channels.Channel
 
   _dispose() {
     // Clean up from parent and connection.
-    if (this._parent)
-      this._parent._objects.delete(this._guid);
+    if (this._parent) this._parent._objects.delete(this._guid);
     this._connection._objects.delete(this._guid);
 
     // Dispose all children.
-    for (const object of [...this._objects.values()])
-      object._dispose();
+    for (const object of [...this._objects.values()]) object._dispose();
     this._objects.clear();
   }
 
   _debugScopeState(): any {
     return {
       _guid: this._guid,
-      objects: Array.from(this._objects.values()).map(o => o._debugScopeState()),
+      objects: Array.from(this._objects.values()).map((o) => o._debugScopeState()),
     };
   }
 
@@ -84,12 +90,23 @@ export abstract class ChannelOwner<T extends channels.Channel = channels.Channel
           const validator = scheme[paramsName(this._type, prop)];
           if (validator) {
             return (params: any) => {
-              return this._wrapApiCall(apiZone => {
-                const { stackTrace, csi, callCookie } = apiZone.reported ? { csi: undefined, callCookie: undefined, stackTrace: null } : apiZone;
+              return this._wrapApiCall((apiZone) => {
+                const { stackTrace, csi, callCookie } = apiZone.reported
+                  ? { csi: undefined, callCookie: undefined, stackTrace: null }
+                  : apiZone;
                 apiZone.reported = true;
                 if (csi && stackTrace && stackTrace.apiName)
-                  csi.onApiCallBegin(renderCallWithParams(stackTrace.apiName, params), stackTrace, callCookie);
-                return this._connection.sendMessageToServer(this, prop, validator(params, ''), stackTrace);
+                  csi.onApiCallBegin(
+                    renderCallWithParams(stackTrace.apiName, params),
+                    stackTrace,
+                    callCookie,
+                  );
+                return this._connection.sendMessageToServer(
+                  this,
+                  prop,
+                  validator(params, ''),
+                  stackTrace,
+                );
               });
             };
           }
@@ -105,12 +122,10 @@ export abstract class ChannelOwner<T extends channels.Channel = channels.Channel
     const logger = this._logger;
     const stack = captureRawStack();
     const apiZone = zones.zoneData<ApiZone>('apiZone', stack);
-    if (apiZone)
-      return func(apiZone);
+    if (apiZone) return func(apiZone);
 
     const stackTrace = captureStackTrace(stack);
-    if (isInternal)
-      delete stackTrace.apiName;
+    if (isInternal) delete stackTrace.apiName;
     const csi = isInternal ? undefined : this._instrumentation;
     const callCookie: any = {};
 
@@ -126,7 +141,8 @@ export abstract class ChannelOwner<T extends channels.Channel = channels.Channel
       logApiCall(logger, `<= ${apiName} succeeded`, isInternal);
       return result;
     } catch (e) {
-      const innerError = ((process.env.PWDEBUGIMPL || isUnderTest()) && e.stack) ? '\n<inner error>\n' + e.stack : '';
+      const innerError =
+        (process.env.PWDEBUGIMPL || isUnderTest()) && e.stack ? '\n<inner error>\n' + e.stack : '';
       e.message = apiName + ': ' + e.message;
       e.stack = e.message + '\n' + frameTexts.join('\n') + innerError;
       csi?.onApiCallEnd(callCookie, e);
@@ -148,8 +164,7 @@ export abstract class ChannelOwner<T extends channels.Channel = channels.Channel
 }
 
 function logApiCall(logger: Logger | undefined, message: string, isNested: boolean) {
-  if (isNested)
-    return;
+  if (isNested) return;
   if (logger && logger.isEnabled('api', 'info'))
     logger.log('api', 'info', message, [], { color: 'cyan' });
   debugLogger.log('api', message);
@@ -164,8 +179,7 @@ export function renderCallWithParams(apiName: string, params: any) {
   const paramsArray = [];
   if (params) {
     for (const name of paramsToRender) {
-      if (params[name])
-        paramsArray.push(params[name]);
+      if (params[name]) paramsArray.push(params[name]);
     }
   }
   const paramsText = paramsArray.length ? '(' + paramsArray.join(', ') + ')' : '';
